requirements.txt
MainContentExtractor>=0.0.4
beautifulsoup4>=4.12.3
httpx>=0.27.2
langchain==0.3.14
langchain-openai==0.3.1
langchain-anthropic==0.3.3
langchain-ollama==0.2.2
langchain-fireworks>=0.2.6
langchain-aws>=0.2.11
langchain-google-genai==2.0.8
pydantic>=2.10.4
python-dotenv>=1.0.1
requests>=2.32.3
posthog>=3.7.0
playwright>=1.49.0
lmnr[langchain]>=0.4.53
